fix: base_install fixes

+++ base_install.sh
@@ -4,15 +4,18 @@
 #
 # curl -fsSL https://raw.githubusercontent.com/lcontreras21/.dotfiles/refs/heads/main/base_install.sh | bash
 
+if command -v apt &> /dev/null; then
+    apt-get update
+    apt-get install git
+fi
 
 git clone https://github.com/lcontreras21/.dotfiles.git --depth 1 /tmp/.dotfiles
 cd /tmp/.dotfiles
 
-if command -v apt &> /dev/null; then
-    apt-get update
+if [ -d "~/.oh-my-zsh" ]; then
+    rm -r ~/.oh-my-zsh/
 fi
 
-rm -r ~/.oh-my-zsh/
 ./install.sh zsh
 cp configs/zsh/.* ~/
